requirements.txt
Flask==2.0.1
pymongo==3.12.0
Werkzeug==2.0.1
python-dotenv==0.19.0
bson==0.5.10
Flask-PyMongo==2.3.0
Flask-Cors==3.0.10
bcrypt==3.2.0
python-dateutil==2.8.2
certifi==2023.7.22 